+++ core/src/main/java/com/asml/apa/wta/core/config/RuntimeConfig.java
@@ -28,4 +28,6 @@ public class RuntimeConfig {
 
   @Builder.Default
   private String logLevel = "ERROR";
+
+  private String outputPath;
 }